RAD-43: Add screen to view messages
Added screen to view messages using a Fragment and the RecycleView which
is a ListView that separates the view from the data.

Edited the main activity (activity_subscribe) to display messages as the
initial view because the content displaying the connection information
will now be displayed in settings.

Note: no work has been done yet on receiving a message. That will be
tackled next. This is just UI work.

https://www.sitepoint.com/mastering-complex-lists-with-the-android-recyclerview/