RAD-45: Improve offline icon hide/show
Improved the offline icon hide/show functionality so that if the app
is relaunched after receiving a message, it will display the correct
status.

Also changed from using getView to findItem which is more specific for
menus.

Allow for situations where the MQTT client updates the connection status
before the menu has been created.